+++ sentry-rails/examples/rails-6.0/app/controllers/welcome_controller.rb
@@ -5,6 +5,9 @@ class WelcomeController < ApplicationController
     1 / 0
   end
 
+  def appearance
+  end
+
   def view_error
   end