+++ sentry-rails/examples/rails-6.0/config/routes.rb
@@ -3,6 +3,7 @@ Rails.application.routes.draw do
   get '500', :to => 'welcome#report_demo'
   root to: "welcome#index"
 
+  get 'appearance', to: 'welcome#appearance'
   get 'view_error', to: 'welcome#view_error'
   get 'worker_error', to: 'welcome#worker_error'
   get 'job_error', to: 'welcome#job_error'